+++ devilry/devilry_group/views/download_files/backends.py
@@ -5,21 +5,24 @@ import os
 from devilry.devilry_compressionutil.backends import backends_base
 
 
-class DevilryGroupZipBackend(backends_base.PythonZipFileBackend):
+class DevilryGroupZipBackend(backends_base.StreamZipBackend):
     """
     Implementation of backend used by devilry group.
     """
     backend_id = 'devilry_group_local'
+    save_to_disk = False
 
     def __init__(self, **kwargs):
         super(DevilryGroupZipBackend, self).__init__(**kwargs)
 
     @classmethod
     def delete_archive(cls, full_path):
-        if not os.path.exists(full_path):
-            return False
-        try:
-            os.remove(full_path)
-        except OSError:
-            return False
-        return True
+        if cls.save_to_disk:
+            if not os.path.exists(full_path):
+                return False
+            try:
+                os.remove(full_path)
+            except OSError:
+                return False
+            return True
+        return False
